+++ src/main/java/com/spring/project/event/dao/EventDAO.java
@@ -65,7 +65,7 @@ public interface EventDAO {
 
 	public List<LodgingDTO> houseList(Map<String, Integer> page);
 
-	public int average(int lod_id);
+	public Double average(int lod_id);
